Changed default integer type to support more than 2^31-1 grid points
Changed the default integer type for grid sizes from int to a 64bit
signed integer type on 64bit systems. This enables using large grids
with more than 2^31 grid points. For maintainability a new typedef
meep::integer is introduced in meep/meeptyes.hpp and is set to
ptrdiff_t. This will ensure the proper sizes on both 32bit and 64bit
systems.

The changes were made according to the following procedure:
- the type of meep::grid_volume::the_ntot was updated to meep::integer
- evrything was recompiled with -Wconversion under gcc4.8.2 on a 64bit
system
- iteratively all places with an implicit conversions from meep::integer
to a smaller interger type where updated until all warnings disappeared
- loop variables were updated from int to meep::integer when a
comparison against meep::integer was involved in the iteration
- all casts to int were checked and corrected where necessary
- additional casts were introduced to silence conversion warnings fron
or to double
- print format specifiers were chenge from %d to %ld with additional
casts to long int to avoid warnings on 32bit systems
- MPI communication wrappers were updated to accept meep::integer as
message size. As MPI specifies only int size messages (and the messages
hopefully will not exceed this limit), internally the size is just
downcasted to int. However, checks of the message size were added before
downcasting to abort when the message size does not allow a safe cast.
This at least ensures a proper error message when the supported message
size is exceeded.

+++ src/casimir.cpp
@@ -298,7 +298,9 @@ complex<double> fields::casimir_stress_dct_integral(direction dforce,
     loop_in_chunks(stress_chunkloop, &data, where, c); 
 
   data.sum = sum_to_all(data.sum);
-  return coefficient * complex<double>(real(data.sum), imag(data.sum));
+  return coefficient * complex<double>(
+		  static_cast<double>(real(data.sum)),
+		  static_cast<double>(imag(data.sum)));
 }
 
   /* Similar to make_g above, but now air/metal systems
